% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_visualize.R
\name{annotation_sankey_plot}
\alias{annotation_sankey_plot}
\title{Annotation Sankey Plot}
\usage{
annotation_sankey_plot(
  object,
  annotations,
  title = "Annotation Sankey Plot",
  nodes = NULL,
  colors = NULL,
  thickness = 15,
  pad = 15,
  font_size = 10
)
}
\arguments{
\item{object}{A Seurat object or a data frame (\code{suerat_obj@meta.data})
with the annotations to visualize}

\item{annotations}{A vector of annotation names to visualize. The order will
dictate the flow in the Sankey plot}

\item{title}{The title of the Sankey plot, default is 'Annotation Sankey Plot'}

\item{nodes}{A vector of unique node/group names to use in the Sankey
plot. If NULL, all unique groups in the annotations will be used}

\item{colors}{A named vector of colors to use for the annotations. The names
should match the annotation group names and should have exactly
the same length as the unique groups in the annotations}

\item{thickness}{The thickness of the Sankey plot, default is 15}

\item{pad}{The padding between the nodes, default is 15}

\item{font_size}{The font size of the text, default is 10}
}
\value{
a plotly object with the Sankey plot
}
\description{
This function generates a Sankey plot for visualizing the
             transitions and overlap between different annotations. The
             Sankey plot is useful for visualizing the flow of cells
             between different levels of annotations.
}
