% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scoring_rnaseq.R
\name{calc_module_scores}
\alias{calc_module_scores}
\title{Calculate gene module score}
\usage{
calc_module_scores(
  data,
  modules,
  expr_normalized = TRUE,
  pool = NULL,
  nbin = 24,
  ctrl = 100,
  seed = 123
)
}
\arguments{
\item{data}{A matrix of gene expression data with genes in rows
and samples in columns. Rownames should be genes and column
names should be samples. Must be a numeric matrix.}

\item{modules}{A list of gene modules or a path to a GMT file or a \code{GeneSetCollection} object.
Each element of the list should be a character vector of gene names.
The names of the list should be the module names.}

\item{expr_normalized}{A logical value indicating whether the data is
normalized. If \code{TRUE}, the function will assume
that the data is already normalized by CPM or TPM or FPKM.}

\item{pool}{A character vector of gene names to be used as background
for the gene module scoring. If \code{NULL}, all genes in
the data will be used as background.}

\item{nbin}{Number of bins to divide genes into based on expression levels.
Default is 24.}

\item{ctrl}{Number of control genes to sample per gene in the gene set.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A data frame with rows as samples and columns as module scores.
}
\description{
Calculate gene module score for each sample using
             average expression of genes in the module with
             background correction. This function is inspired by
             Seurat's \code{AddModuleScore} function and is adapted
             to be used with RNA-seq data.
}
