% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scoring_rnaseq.R
\name{calc_module_scores_ssgsea}
\alias{calc_module_scores_ssgsea}
\title{Calculate Gene module enrichment using ssGSEA}
\usage{
calc_module_scores_ssgsea(
  data,
  modules,
  expr_normalized = TRUE,
  minSize = 1,
  maxSize = Inf,
  alpha = 0.75,
  normalize = TRUE
)
}
\arguments{
\item{data}{A matrix or \code{ExpressionSet} of gene expression data with genes in rows
and samples in columns. Rownames should be genes and column
names should be samples. Must be a numeric matrix or an \code{ExpressionSet}.}

\item{modules}{A list of gene modules or a path to a GMT file or a \code{GeneSetCollection} object.
Each element of the list should be a character vector of gene names.
The names of the list should be the module names.}

\item{expr_normalized}{A logical value indicating whether the data is
normalized. If \code{TRUE}, the function will assume
that the data is already normalized by CPM or TPM or FPKM.}

\item{minSize}{Minimum size of gene set to be considered for ssGSEA. Default is 1.}

\item{maxSize}{Maximum size of gene set to be considered for ssGSEA. Default is Inf.}

\item{alpha}{The exponent defining the weight of the tail in the random walk
performed by the ssGSEA. See \code{GSVA::ssgseaParam} for more details.}

\item{normalize}{A logical value indicating whether to normalize the enrichment scores.}
}
\value{
A data frame with rows as samples and columns as module scores.
}
\description{
Calculate gene module enrichment scores using single-sample
             gene set enrichment analysis (ssGSEA) method. This function
             utilizes the GSVA package to calculate enrichment scores for
             gene modules in each sample.
}
