% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_visualize.R
\name{de_dot_plot}
\alias{de_dot_plot}
\title{Dot Plot for comparisons with logFC and p-values}
\usage{
de_dot_plot(
  de_results,
  features,
  feature_var = "gene",
  logFC_var = "avg_log2FC",
  p_val_var = "p_val_adj",
  comparison_var = "comparison",
  logFC_cutoff = 0,
  p_val_cutoff = 0.05,
  palette = NULL,
  down_color = "darkgreen",
  mid_color = "white",
  up_color = "darkslateblue",
  nsig_color = "grey70",
  sig_only = FALSE,
  invert = FALSE
)
}
\arguments{
\item{de_results}{A data frame with the differential expression results, must
contain the columns for "feature", "avg_logFC", "p_val_adj",
and "comparison"}

\item{features}{A vector of feature names to plot}

\item{feature_var}{The name of the feature variable in the de_results data
frame, default is "gene"}

\item{logFC_var}{The name of the logFC variable in the de_results data frame,
default is "avg_logFC"}

\item{p_val_var}{The name of the p-value variable in the de_results data frame,
default is "p_val_adj"}

\item{comparison_var}{The name of the comparison variable in the de_results
data frame, default is "comparison"}

\item{logFC_cutoff}{The logFC cutoff for filtering the features, default is 0}

\item{p_val_cutoff}{The p-value cutoff for filtering the features, default is
0.05}

\item{palette}{The color palette to use for the dot plot for representing the
logFC values, default is "NULL". If NULL, down_color, mid_color,
and up_color will be used. If not NULL, the palette will be used
to represent the logFC values and the down_color, mid_color, and
up_color will be ignored}

\item{down_color}{The color for down-regulated features, default is "darkgreen"}

\item{mid_color}{The color for 0 logFC features, default is "white"}

\item{up_color}{The color for up-regulated features, default is "darkslateblue"}

\item{nsig_color}{The color for non-significant features, default is "grey70"}

\item{sig_only}{Whether to plot only significant features, default is FALSE,
which plots all features provided. If TRUE, the plot will
only show features-comparisons with p-values < p_val_cutoff
and logFC > logFC_cutoff}

\item{invert}{Whether to invert the plot, default is FALSE. By default, the
x axis is the features and the y axis is the comparisons. If
TRUE, the x axis is the comparisons and the y axis is the
features}
}
\value{
a ggplot object with the dot plot
}
\description{
This function generates a dot plot for visualizing comparisons
             between multiple pairs of groups. The dot plot shows the log
             fold change (logFC) and p-values for each comparison for each
             feature. The dot plot is useful for visualizing differential
             expression analysis results. If a gene in a given comparison has
             a significant p-value, the dot is colored based on the logFC
             value. The size of the dot is proportional to the -log10 of the
             p-value. If not significant, the dot is colored in grey.
}
