% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_visualize.R
\name{do_FeaturePlot_arrow}
\alias{do_FeaturePlot_arrow}
\title{do_FeaturePlot with arrows on the bottom left corner}
\usage{
do_FeaturePlot_arrow(
  object,
  features,
  assay = NULL,
  reduction = "umap",
  slot = "data",
  order = TRUE,
  dims = NULL,
  pt.size = 0.5,
  legend.position = "right",
  arrow_length = 0.1,
  arrow_x_adjust_ratio = 1,
  arrow_y_adjust_ratio = 1,
  x_coord_adjust = 0,
  y_coord_adjust = 0,
  arrow_size = 0.2,
  arrow_width = 0.9,
  arrow_text_color = "black",
  arrow_type = "closed",
  text_font_size = 4,
  relative_text_dst = 0.15,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{A vector of feature names to plot}

\item{assay}{The assay to use for plotting, default is the active assay}

\item{reduction}{The reduction method used to generate the plot, default is
"umap".}

\item{slot}{The slot to use for the reduction, default is "data"}

\item{order}{Whether to order the cells based on feature expression. Default
is TRUE}

\item{dims}{A vector of dimension names, overwrites the default dimension
names assumed from reduction method. (Overwrite the dims argument
in the SCpubr::do_FeaturePlot function)}

\item{pt.size}{The size of the points, default is 0.5}

\item{legend.position}{The position of the legend, default is 'right'}

\item{arrow_length}{The size of the arrow, default is 0.15 (15% of the plot)}

\item{arrow_x_adjust_ratio}{The x-axis adjustment ratio for the arrow,}

\item{arrow_y_adjust_ratio}{The y-axis adjustment ratio for the arrow,}

\item{x_coord_adjust}{Horizontal adjustment of the arrow origin, default is 0}

\item{y_coord_adjust}{Vertical adjustment of the arrow origin, default is 0}

\item{arrow_size}{The size of the arrow head, default is 0.2 cm}

\item{arrow_width}{The width of the arrow, default is 0.9}

\item{arrow_text_color}{The color of the arrow and label, default is "black"}

\item{arrow_type}{The type of the arrow, default is "closed"}

\item{text_font_size}{The font size of the text, default is 4}

\item{relative_text_dst}{The relative distance of the text from the arrow,}

\item{...}{Additional arguments to pass to the SCpubr::do_FeaturePlot function}
}
\value{
a ggplot object with the dimension arrows added
}
\description{
This function is a further implementation of the do_FeaturePlot
            function from the SCpubr package. It generates a ggplot object
           with the dimension arrows added to the bottom left corner of
           the plot. The function is useful for visualizing single cell
          data with dimension reduction methods such as UMAP, t-SNE, PCA,
         and spatial coordinates.
}
