% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rnaseq.R
\name{filter_de_genes}
\alias{filter_de_genes}
\title{Filter genes based on differential expression analysis results}
\usage{
filter_de_genes(
  de_results,
  log2fc_min = 0,
  pvalue_max = 0.05,
  de_method = "edgeR",
  pvalue_col = NULL,
  log2fc_col = NULL
)
}
\arguments{
\item{de_results}{A data frame containing the results of differential
expression analysis. The data frame should have at least
a column for log2 fold change and a column for (adjusted)
p-value.}

\item{log2fc_min}{The minimum log2 fold change to filter genes. Default is
0, which means no filtering based on log2 fold change.}

\item{pvalue_max}{The maximum adjusted p-value to filter genes. Default is
0.05.}

\item{de_method}{The method used for differential expression analysis.
Must be one of "edgeR" or "DESeq2". Default is "edgeR".
Use this argument to automatically set the column names
for log2 fold change and adjusted P-value. If provided value
is not "edgeR" or "DESeq2", the user must provide both the
column names for log2 fold change and P-value using the
\code{log2fc_col} and \code{pvalue_col} arguments,
respectively.}

\item{pvalue_col}{The name of the column in \code{de_results} that contains
the p-values. If not NULL, overrides the default column
name based on \code{de_method}.}

\item{log2fc_col}{The name of the column in \code{de_results} that contains
the log2 fold changes.}
}
\value{
A list of two data frames:
       \itemize{
         \item{up_reg}{Data frame containing up-regulated genes}
         \item{down_reg}{Data frame containing down-regulated genes}
       }
}
\description{
This function filters genes based on the results of differential
             expression analysis for RNA-seq data. The user can specify the
             minimum log2 fold change and maximum adjusted p-value to filter
             the genes.
}
