% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rnaseq.R
\name{filter_low_expression_genes}
\alias{filter_low_expression_genes}
\title{Filter out genes with low expression based on minimum group size}
\usage{
filter_low_expression_genes(raw_counts, min_group_size = 4)
}
\arguments{
\item{raw_counts}{A data frame containing the raw counts of genes. The data
frame should have samples as columns and genes as rows.
All values should be non-negative integers.}

\item{min_group_size}{The minimum number of samples in which the gene should
be expressed. Default is 4.}
}
\value{
A data frame containing the filtered raw counts
}
\description{
This function filters out genes with low expression based on
             the minimum group size. The user can specify the minimum number
             of samples in which the gene should be expressed. Essentially,
             a gene is kept only if it has at least 1 count per million
             (CPM) in at least \code{min_group_size} samples.
}
