% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_visualize.R
\name{format_sc_plots}
\alias{format_sc_plots}
\title{Format Dimension Plots (from Seurat, scCustomize, SCpubr etc.)}
\usage{
format_sc_plots(
  plots,
  ncol = NULL,
  nrow = NULL,
  byrow = TRUE,
  add_arrow = TRUE,
  bottom_left_only = FALSE,
  remove_axes = TRUE,
  remove_legend = FALSE,
  remove_grid = FALSE,
  restore_legend = FALSE,
  legend_position = NULL,
  legend_title = NULL,
  legend_title_size = 12,
  legend_title_bold = FALSE,
  reduction = "umap",
  dims = NULL,
  arrow_length = 2,
  x_coord_adjust = 0.25,
  y_coord_adjust = x_coord_adjust,
  arrow_size = 0.25,
  arrow_width = 0.2,
  arrow_line_width = 3,
  arrow_color = "black",
  text_font_size = 12,
  text_dst = 0.3
)
}
\arguments{
\item{plots}{A list of ggplot objects or a patchwork object}

\item{ncol}{Number of columns in the plot grid, default is NULL. Prioritized if
nrow is also set.}

\item{nrow}{Number of rows in the plot grid, default is NULL}

\item{byrow}{Whether to fill the plot grid by row, default is TRUE, consistent
with patchwork's default behavior. If FALSE, the plots will be
filled by column.}

\item{add_arrow}{Whether to add an arrow to replace the axes, default is TRUE.
This will override the remove_axes argument and remove the axes
from all plots}

\item{bottom_left_only}{Whether to add the arrow only to the plot at the bottom
left of the entire grid, default is FALSE. Only valid
if the arrangement of the plots are defined, either by
setting ncol or nrow, or that patchwork already defines
the layout. Only used when add_arrow is TRUE}

\item{remove_axes}{Whether to remove the axes, default is TRUE}

\item{remove_legend}{Whether to remove the legend, default is FALSE}

\item{remove_grid}{Whether to remove the grid, default is FALSE}

\item{restore_legend}{Whether to restore the legend, default is FALSE}

\item{legend_position}{The position of the legend, default is NULL (meaning no change)}

\item{legend_title}{The title of the legend, default is NULL}

\item{legend_title_size}{The size of the legend title, default is 12}

\item{legend_title_bold}{Whether to bold the legend title, default is FALSE}

\item{reduction}{The reduction method used to generate the plots, default is
"umap", allowed values are "umap", "tsne", "pca", "spatial".
Used only if add_arrow is TRUE}

\item{dims}{A vector of dimension names, overwrites the default dimension.
Used only if add_arrow is TRUE}

\item{arrow_length}{The size of the arrow, default is 2 (cm)
Used only if add_arrow is TRUE}

\item{x_coord_adjust}{Horizontal adjustment of the arrow origin, default is 0.25 (cm)
Used only if add_arrow is TRUE}

\item{y_coord_adjust}{Vertical adjustment of the arrow origin, default is the same as x_coord_adjust
Used only if add_arrow is TRUE}

\item{arrow_size}{The size of the arrow head, default is 0.25 cm
Used only if add_arrow is TRUE}

\item{arrow_width}{The width of the arrow, default is 0.2 cm
Used only if add_arrow is TRUE}

\item{arrow_line_width}{The width of the arrow line, default is 3
Used only if add_arrow is TRUE}

\item{arrow_color}{The color of the arrow and label, default is "black"
Used only if add_arrow is TRUE}

\item{text_font_size}{The font size of the text, default is 12
Used only if add_arrow is TRUE}

\item{text_dst}{The distance of the text from the arrow in cm, default is 0.3
Used only if add_arrow is TRUE}
}
\value{
a patchwork object with the formatted plots
}
\description{
This function operates on a list of
             ggplot objects or a patchwork object and formats them to have a
             consistent theme. The function performs group operations to
             format the plots. The default behavior is to remove the axes and
             add an arrow to the bottom left corner of each individual plot.
}
