% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rnaseq.R
\name{generate_de_rank_file}
\alias{generate_de_rank_file}
\title{Generate rank file for GSEA}
\usage{
generate_de_rank_file(
  de_results,
  save_file = NULL,
  de_method = "edgeR",
  pvalue_col = NULL,
  log2fc_col = NULL
)
}
\arguments{
\item{de_results}{A data frame containing the results of differential
expression analysis. The data frame should have at least
a column for log2 fold change and a column for (adjusted)
p-value.}

\item{save_file}{The name of the file to save the rank file. Default is NULL.
If NULL, the rank file will not be saved to disk.}

\item{de_method}{The method used for differential expression analysis.
Must be one of "edgeR" or "DESeq2". Default is "edgeR".
Use this argument to automatically set the column names
for log2 fold change and adjusted P-value. If provided value
is not "edgeR" or "DESeq2", the user must provide both the
column names for log2 fold change and P-value using the
\code{log2fc_col} and \code{pvalue_col} arguments,
respectively.}

\item{pvalue_col}{The name of the column in \code{de_results} that contains
the p-values. If not NULL, overrides the default column
name based on \code{de_method}.}

\item{log2fc_col}{The name of the column in \code{de_results} that contains
the log2 fold changes.}
}
\value{
A data frame with the addition of the following columns:
         \item{piValue}{The pi-value of the gene}
         \item{rank}{The rank of the gene based on the pi-value}
        And if \code{save_file} is not NULL, a rank file will be saved to disk.
}
\description{
This function generates a rank file for Gene Set Enrichment
             Analysis (GSEA) based on the differential expression analysis
             results. The rank file contains the gene names and their ranks,
             which is the pi-value. Higher pi-values indicate higher
             rank.
}
