% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_utils.R
\name{get_meta_pivot}
\alias{get_meta_pivot}
\title{Generate a pivot table from cell metadata}
\usage{
get_meta_pivot(
  row_metadata,
  col_metadata,
  type = c("percent", "count"),
  margin = c("row", "column", "total")
)
}
\arguments{
\item{row_metadata}{A vector of cell metadata to be used as rows, often a
metadata that represents cell clusters or cell types}

\item{col_metadata}{A vector of cell metadata to be used as columns, must be
the same length as row_metadata}

\item{type}{The type of pivot table to generate, either "percent" or "count"}

\item{margin}{The margin to calculate the percentage, either "row", "column"
or "total", used only when type = "percent"}
}
\value{
a data frame with the pivot table
}
\description{
This function generates a pivot table from cell metadata and
             returns a data frame with the two cell metadata as rows and
             columns, with the values being either the count of cells or
             the percentage of cells (by rows or columns or total).
}
