% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ggplotColours}
\alias{ggplotColours}
\title{Getting default ggplot2 colors}
\usage{
ggplotColours(n = 6, h = c(0, 360) + 15, showColor = FALSE)
}
\arguments{
\item{n}{Number of categories}

\item{h}{Hue range}

\item{showColor}{Show the colors, default is FALSE}
}
\value{
A vector of colors (hexadecimal)
}
\description{
Since Seurat DimPlot uses ggplot2 default colors when no palette is
specified, different categories will result in lack of consistency in
colors (ggplot2 ensure the first the last color are the same, but the
middle colors are different depending on the number of categories). This
function returns the default ggplot2 colors for a given number of categories.
This function is directly taken and modified from the following StackOverflow post:
https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
\references{
https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
