% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\name{keywords_to_info}
\alias{keywords_to_info}
\title{Extracting sample info based on regex keywords}
\usage{
keywords_to_info(sample, keywords, info)
}
\arguments{
\item{sample}{A string containing the sample info, must be a single string}

\item{keywords}{A vector of regex keywords to search for}

\item{info}{A vector of info represented by the keywords, must be the same
length as keywords}
}
\value{
the info extracted from the sample based on the keywords found
}
\description{
Extracting sample info based on regex keywords found in a
             string. If multiple keywords are found, the first keyword
             found will be used to extract the info.
}
\examples{
# Extracting sample info based on regex keywords
sample <- "Sample_2wks_1"
keywords <- c("[0-9]wks", "p7")
info <- c("postnatal_2wks", "postnatal_day_7")

# Extracting the stage info
keywords_to_info(sample, keywords, info)

}
