% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_visualize.R
\name{manipulate_sc_plot}
\alias{manipulate_sc_plot}
\title{Helper to manipulate ggplots using Seurat's patchwork functions}
\usage{
manipulate_sc_plot(
  plot,
  remove_axes = FALSE,
  remove_legend = FALSE,
  remove_grid = FALSE,
  restore_legend = FALSE,
  legend_position = NULL,
  legend_title = NULL,
  legend_title_size = 12,
  legend_title_bold = FALSE
)
}
\arguments{
\item{plot}{A single ggplot object}

\item{remove_axes}{Whether to remove the axes, default is FALSE}

\item{remove_legend}{Whether to remove the legend, default is FALSE}

\item{remove_grid}{Whether to remove the grid, default is FALSE}

\item{restore_legend}{Whether to restore the legend, default is FALSE}

\item{legend_position}{The position of the legend, default is NULL (meaning no change)}

\item{legend_title}{The title of the legend, default is NULL}

\item{legend_title_size}{The size of the legend title, default is 12}

\item{legend_title_bold}{Whether to bold the legend title, default is FALSE}
}
\value{
a ggplot object with the axes, legend, and grid manipulated
}
\description{
Only valid for individual ggplot objects
}
