% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mix_colors}
\alias{mix_colors}
\title{Mix colors based on weights of each color}
\usage{
mix_colors(colors, weights = rep(1, length(colors)))
}
\arguments{
\item{colors}{A vector of colors (hexadecimal)}

\item{weights}{A vector of weights (numeric)}
}
\value{
A mixed color (hexadecimal)
}
\description{
This function mixes colors based on the weights of each color.
}
