% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_atac.R
\name{peak_calc_frip}
\alias{peak_calc_frip}
\title{Calculate FRiP score for peaks}
\usage{
peak_calc_frip(
  peaks,
  peak_names = NULL,
  fragment_file,
  threads = 8,
  is_indexed = NULL
)
}
\arguments{
\item{peaks}{A list of GRanges objects, each representing peaks identified
by a different peak calling method or sample.}

\item{peak_names}{A character vector of names corresponding to the peaks in
 `peaks`. If not provided, names will be extracted from
`peaks` names. Default is `NULL`.}

\item{fragment_file}{A character string specifying the path to the fragment file
(in BED format) containing the reads. This file should
be deduplicated and sorted. This file can be either
Tabix indexed or not. However, for large files, it is
recommended to use a Tabix indexed file for faster access.}

\item{threads}{An integer specifying the number of threads to use for parallel processing.
Default is 8.}

\item{is_indexed}{A logical value indicating whether the fragment file is Tabix indexed.
Default is to detect automatically (NULL).}
}
\value{
A data frame containing the FRiP scores for each peak set.
}
\description{
This function calculates the FRiP (Fraction of Reads in Peaks) score
for a given set of peaks. It computes the fraction of reads that overlap with
the peaks and returns a data frame with the FRiP scores for each peak set.
}
