% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_atac.R
\name{peak_category_plot}
\alias{peak_category_plot}
\title{Plotting peak category across one or more peak calling methods/samples}
\usage{
peak_category_plot(
  peak_list,
  peak_names = NULL,
  TxDb,
  palette = "polychrome",
  plot_title = "Peak Category QC"
)
}
\arguments{
\item{peak_list}{A named list of GRanges objects, each representing peaks
identified by a different peak calling method or sample.}

\item{peak_names}{A character vector of names corresponding to the peaks in
`peak_list`. If not provided, names will be extracted from
`peak_list` names. Default is `NULL`.}

\item{TxDb}{A TxDb object containing gene annotations. This is used to
categorize peaks into promoter, intronic, intergenic, and
exonic regions.}

\item{palette}{A valid palette name from "alphabet", "alphabet2", "glasbey", "polychrome", "stepped", and "parade".
Alternatively, a vector of colors can be provided with the same length as unique groups.}

\item{plot_title}{A character string for the title of the plot. Default is
"Peak Category QC".}
}
\value{
A ggplot object representing the bar plot of peak categories.
}
\description{
This function performs quality control on peak categories across
             one or more peak calling methods. It generates a bar plot
             showing the distribution of peaks across different categories.
}
