% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_atac.R
\name{peak_plot_frip}
\alias{peak_plot_frip}
\title{Plot FRiP scores}
\usage{
peak_plot_frip(frip_df, palette = "polychrome", plot_title = "FRiP Score")
}
\arguments{
\item{frip_df}{A data frame containing the FRiP scores for each peak set, the output of `peak_calc_frip`.}

\item{palette}{A valid palette name from "alphabet", "alphabet2", "glasbey", "polychrome", "stepped", and "parade".
Alternatively, a vector of colors can be provided with the same length as unique groups.}

\item{plot_title}{A character string for the title of the plot. Default is "FRiP Score".}
}
\value{
A ggplot object representing the bar plot of FRiP scores.
}
\description{
This function generates a bar plot of FRiP (Fraction of Reads in Peaks)
}
