% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_atac.R
\name{peak_tss_enrichment}
\alias{peak_tss_enrichment}
\title{Peak TSS enrichment}
\usage{
peak_tss_enrichment(
  peaks,
  peak_names = NULL,
  TxDb,
  upstream = 3000,
  downstream = 3000,
  plot_title = "Peak TSS Enrichment"
)
}
\arguments{
\item{peaks}{A list of GRanges objects, each representing peaks identified
by a different peak calling method or sample.}

\item{peak_names}{A character vector of names corresponding to the peaks in
                 `peaks`. If not provided, names will be extracted from
'                `peaks` names. Default is `NULL`.}

\item{TxDb}{A TxDb object containing gene annotations.}

\item{upstream}{A numeric value specifying the upstream region to consider for TSS enrichment.}

\item{downstream}{A numeric value specifying the downstream region to consider for TSS enrichment.}

\item{plot_title}{A character string for the title of the plot. Default is
"Peak TSS Enrichment".}
}
\value{
A ggplot object representing the TSS enrichment plot.
}
\description{
This function calculates the TSS (Transcription Start Site) enrichment
for a given set of peaks. It uses the `ChIPseeker` package to annotate the peaks
and then computes the TSS enrichment.
}
