% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_atac.R
\name{peak_width_distribution}
\alias{peak_width_distribution}
\title{Peak width distribution}
\usage{
peak_width_distribution(
  peaks,
  peak_names = NULL,
  max_length = 10000,
  plot_title = "Peak Width Distribution"
)
}
\arguments{
\item{peaks}{A list of GRanges objects, each representing peaks identified
by a different peak calling method or sample.}

\item{peak_names}{A character vector of names corresponding to the peaks in
`peaks`. If not provided, names will be extracted from
`peaks` names. Default is `NULL`.}

\item{max_length}{An integer specifying the maximum length of the peaks to consider.}

\item{plot_title}{A character string for the title of the plot. Default is}
}
\value{
A patchwork object containing a histogram of peak widths with a density
        curve overlay and a boxplot showing the distribution of peak widths
        across different peak sets.
}
\description{
This function calculates the distribution of peak widths for a
             given set of peaks. It generates a histogram with a density
             curve overlay to visualize the distribution of peak widths. A
             boxplot is also generated to show the distribution of peak
             widths across different peak sets.
}
