% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_visualize.R
\name{plot_meta_scatter}
\alias{plot_meta_scatter}
\title{Metadata Scatter Plot}
\usage{
plot_meta_scatter(
  object,
  group.by,
  split.by,
  cells.use = NULL,
  group.use = NULL,
  x_condition = NULL,
  y_condition = NULL,
  x_label = NULL,
  y_label = NULL,
  mode = c("prop", "count"),
  percentage = TRUE,
  colors = NULL,
  point_size = 4,
  point_alpha = 0.8,
  point_border = TRUE,
  point_border_stroke = 1,
  add_diagonal = TRUE,
  diagonal_slope = 1,
  diagonal_intercept = 0,
  diagonal_color = "grey60",
  diagonal_linetype = "11",
  diagonal_size = 0.8,
  label = TRUE,
  label_text_size = 4,
  legend_position = "none",
  legend_title = group.by,
  remove_zero = TRUE
)
}
\arguments{
\item{object}{A Seurat object or a data.frame containing the metadata
(e.g., cell type annotations). If a dataframe is provided,
it must be in long format with columns for each variable, e.g.,
`seurat_obj@meta.data`.}

\item{group.by}{Character vector specifying the column name(s) containing the
grouping variable(s) (e.g., cell types, clusters).}

\item{split.by}{Character string specifying the column name containing the
condition variable (e.g., treatment, genotype).}

\item{cells.use}{Optional character vector specifying the cells to include in the
analysis. If NULL (default), all cells are included. Can be
`WhichCells(seurat_obj, idents = c("ident1", "ident2"))` or
any other vector of cell names.}

\item{group.use}{Optional character vector specifying the groups to include
in the plot. If NULL (default), all groups are included.
When plotting proportions, setting this parameter does NOT
affect the proportion calculations, which are always based
on all groups present in the data. This will only filter the
points shown in the plot.}

\item{x_condition}{The name of the condition to plot on the x-axis. Must be
 one of the unique values in the split.by column. Optional
(default NULL) if there are only two unique conditions.}

\item{y_condition}{The name of the condition to plot on the y-axis. Must be
 one of the unique values in the split.by column. Optional
(default NULL) if there are only two unique conditions.}

\item{x_label}{The label for the x-axis. If NULL (default), the x_condition
will be used as the label.}

\item{y_label}{The label for the y-axis. If NULL (default), the y_condition
will be used as the label.}

\item{mode}{Either "prop" or "count" to determine whether to plot the
proportion or count of cells for each group in the specified
conditions.}

\item{percentage}{Logical, whether to convert proportions to percentages
(default FALSE).}

\item{colors}{A named vector of colors to use for the groups. The names
should match the group names and should have exactly the same
length as the number of unique groups.}

\item{point_size}{The size of the points in the scatter plot, default is 3.}

\item{point_alpha}{The transparency of the points in the scatter plot,
default is 0.8.}

\item{point_border}{Whether to add a black border around the points,
default is TRUE.}

\item{point_border_stroke}{The thickness of the point border,
default is 0.5.}

\item{add_diagonal}{Logical, whether to add a diagonal reference line
(default TRUE).}

\item{label}{Logical, whether to add labels to the points (default TRUE).}

\item{label_text_size}{The size of the point labels, default is 3.5.}

\item{legend_position}{The position of the legend, default is 'none'.}

\item{legend_title}{The title of the legend, default is the value of group.by.}

\item{remove_zero}{Logical, whether to remove groups with zero counts in both
conditions (default TRUE).}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
This function generates a scatter plot for visualizing the
             relationship of a grouped metadata variable between two conditions.
             Each point represents a group (e.g., cell type) and its position
             is determined by the proportion or count of that group in each
             of two specified conditions. This is useful for comparing how
             the distribution of groups changes between conditions.
}
