% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rnaseq.R
\name{plot_top_pca_loadings}
\alias{plot_top_pca_loadings}
\title{Plotting top PCA loadings}
\usage{
plot_top_pca_loadings(
  pca_obj,
  pcs = 1:3,
  top_n = 20,
  low_color = "darkgreen",
  mid_color = "white",
  high_color = "darkslateblue"
)
}
\arguments{
\item{pca_obj}{A PCA object generated using the \code{prcomp} function.}

\item{pcs}{A vector of integers specifying the principal components to
extract the loadings. Default is to extract PC 1-3.}

\item{top_n}{The number of top genes to extract for each principal component.
Default is 20.}

\item{low_color}{The color for low loadings}

\item{mid_color}{The color for mid loadings}

\item{high_color}{The color for high loadings}
}
\value{
A list, where each element is a PC, and each element is a named vector
        containing the top genes and their loadings.
}
\description{
This function extracts the PCA loadings from a PCA object
             generated using the \code{prcomp} function and plots the top
             genes for each principal component.
}
