% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rnaseq.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot for gene expression analysis}
\usage{
plot_volcano(
  data,
  de_method = "edgeR",
  x = NULL,
  y = NULL,
  label = rownames(data),
  selected_labels = NULL,
  p_cutoff = 0.05,
  fc_cutoff = 0,
  up_color = "firebrick4",
  down_color = "midnightblue",
  non_de_color = "grey",
  up_label = "Up-regulated",
  down_label = "Down-regulated",
  non_de_label = "Non-DE",
  number_label = TRUE,
  num_label_vjust = 0,
  num_label_size = 4.5,
  x_lim = c(min(data[[x]], na.rm = TRUE) - 0.5, max(data[[x]], na.rm = TRUE) + 0.5),
  y_lim = c(0, max(-log10(data[[y]]), na.rm = TRUE) + 1),
  x_lab = bquote(~log[2] ~ "fold change"),
  y_lab = bquote(~-log[10] ~ adj.P),
  axis_lab_size = 12,
  lab_size = 4,
  bold_text = TRUE,
  legend_lab_size = 12,
  legend_lab_position = "top",
  legend_icon_size = 5,
  draw_connectors = TRUE,
  max_overlaps = 10,
  ...
)
}
\arguments{
\item{data}{A data frame containing the gene expression data}

\item{de_method}{The method used for differential expression analysis.
Must be one of "edgeR" or "DESeq2". Default is "edgeR".
Use this argument to automatically set the column names
for log2 fold change and adjusted P-value. If provided value
is not "edgeR" or "DESeq2", the user must provide both the
column names for log2 fold change and P-value using the
\code{x} and \code{y} arguments, respectively.}

\item{x}{The column name of the log2 fold change. If not \code{NULL},
overrides the value obtained from de_method.}

\item{y}{The column name of the P-value for use. If not \code{NULL},
overrides the value obtained from de_method.}

\item{label}{The column name of the gene names}

\item{selected_labels}{A vector of gene names to be highlighted in the plot}

\item{p_cutoff}{The p-value cutoff for significance}

\item{fc_cutoff}{The log2 fold change cutoff for significance}

\item{up_color}{The color for up-regulated genes}

\item{down_color}{The color for down-regulated genes}

\item{non_de_color}{The color for non-differentially expressed genes}

\item{up_label}{The label for up-regulated genes}

\item{down_label}{The label for down-regulated genes}

\item{non_de_label}{The label for non-differentially expressed genes}

\item{number_label}{A logical value indicating whether to display the number
of up and down-regulated genes}

\item{num_label_vjust}{The vertical justification for the number labels}

\item{num_label_size}{The size of the number labels}

\item{x_lim}{The limits for the x-axis}

\item{y_lim}{The limits for the y-axis}

\item{x_lab}{The label for the x-axis}

\item{y_lab}{The label for the y-axis}

\item{axis_lab_size}{The size of the axis labels}

\item{lab_size}{The size of the gene labels}

\item{bold_text}{A logical value indicating whether to bold the text}

\item{legend_lab_size}{The size of the legend labels}

\item{legend_lab_position}{The position of the legend labels}

\item{legend_icon_size}{The size of the legend icons}

\item{draw_connectors}{A logical value indicating whether to draw connectors}

\item{max_overlaps}{The maximum number of overlaps for the gene labels}

\item{...}{Additional arguments to be passed to the plot function}
}
\value{
A volcano plot
}
\description{
Generate a volcano plot for gene expression analysis based on
             the log2 fold change and the -log10 p-value.
}
\details{
This function inherently uses the EnhancedVolcano package to
         generate a volcano plot for gene expression analysis. Some
         aesthetic are modified from the default plot to make it more
         suitable for RNA-seq data analysis. Use the \code{...} to provide
         additional arguments to EnhanceVolcano::EnhancedVolcano.
}
