% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rnaseq.R
\name{plot_volcano2}
\alias{plot_volcano2}
\title{Volcano plot for gene expression analysis}
\usage{
plot_volcano2(
  data,
  de_method = "edgeR",
  x = NULL,
  y = NULL,
  label = "row.names",
  selected_labels = rownames(data),
  draw_connectors = TRUE,
  p_cutoff = 0.05,
  fc_cutoff = 0,
  up_color = "firebrick4",
  down_color = "midnightblue",
  non_de_color = "grey",
  up_label = "Up-regulated",
  down_label = "Down-regulated",
  non_de_label = "Non-DE",
  number_label = TRUE,
  up_arrow_text = up_label,
  down_arrow_text = down_label,
  arrow_text_wrap = 15,
  arrow_y_offset = 0.5,
  arrow_x_offset = 0.05,
  arrow_color = "black",
  arrow_size = 0.2,
  arrow_length = 0.42,
  arrow_text_size = 4,
  text_label_vjust = -0.5,
  num_label_vjust = 1.5,
  num_label_size = 4.5,
  alpha = 0.5,
  pt_size = 2,
  x_lim = NULL,
  y_lim = NULL,
  x_lab = bquote(~log[2] ~ "fold change"),
  y_lab = bquote(~-log[10] ~ adj.P),
  axis_lab_size = 12,
  lab_size = 3.5,
  label_arrow = TRUE,
  label_arrow_type = "closed",
  label_arrow_ends = "last",
  label_force = 10,
  label_point_padding = 1,
  label_box_padding = 0.4,
  label_segment_curvature = 0,
  label_segment_angle = 90,
  label_segment_ncp = 1,
  max_overlaps = 10,
  ...
)
}
\arguments{
\item{data}{A data frame containing the gene expression data}

\item{de_method}{The method used for differential expression analysis.
Must be one of "edgeR", "DESeq2", "limma", or "seurat". Default is "edgeR".
Use this argument to automatically set the column names
for log2 fold change and adjusted P-value. If provided value
is not "edgeR", "DESeq2", "limma", or "seurat", the user must provide both the
column names for log2 fold change and P-value using the
\code{x} and \code{y} arguments, respectively.}

\item{x}{The column name of the log2 fold change. If not \code{NULL},
overrides the value obtained from de_method.}

\item{y}{The column name of the P-value for use. If not \code{NULL},
overrides the value obtained from de_method.}

\item{label}{The column name of the gene names}

\item{selected_labels}{A vector of gene names to be highlighted in the plot}

\item{draw_connectors}{A logical value indicating whether to draw connectors}

\item{p_cutoff}{The p-value cutoff for significance}

\item{fc_cutoff}{The log2 fold change cutoff for significance}

\item{up_color}{The color for up-regulated genes}

\item{down_color}{The color for down-regulated genes}

\item{non_de_color}{The color for non-differentially expressed genes}

\item{up_label}{The label for up-regulated genes}

\item{down_label}{The label for down-regulated genes}

\item{non_de_label}{The label for non-differentially expressed genes}

\item{number_label}{A logical value indicating whether to display the number
of up and down-regulated genes}

\item{up_arrow_text}{The text to display for the up arrow}

\item{down_arrow_text}{The text to display for the down arrow}

\item{arrow_text_wrap}{The maximum number of characters to display on one line}

\item{arrow_y_offset}{The vertical offset for the arrow text}

\item{arrow_x_offset}{The horizontal offset for the arrow text}

\item{arrow_color}{The color of the arrows}

\item{arrow_size}{The size of the arrows}

\item{arrow_length}{The length of the arrows}

\item{arrow_text_size}{The size of the arrow text}

\item{text_label_vjust}{The vertical justification for the text labels}

\item{num_label_vjust}{The vertical justification for the number labels}

\item{num_label_size}{The size of the number labels}

\item{alpha}{The transparency level of the points in the plot}

\item{pt_size}{The size of the points in the plot}

\item{x_lim}{The limits for the x-axis}

\item{y_lim}{The limits for the y-axis}

\item{x_lab}{The label for the x-axis}

\item{y_lab}{The label for the y-axis}

\item{axis_lab_size}{The size of the axis labels}

\item{lab_size}{The size of the gene labels}

\item{label_arrow}{A logical value indicating whether to draw arrow connectors}

\item{label_arrow_type}{The arrow style: "closed", "open"}

\item{label_arrow_ends}{Where the arrow points: "first" (toward label), "last" (toward dot)}

\item{label_force}{The repulsion force for text spreading}

\item{label_point_padding}{The padding to avoid overlapping points}

\item{label_box_padding}{The padding to avoid overlapping text boxes}

\item{label_segment_curvature}{The curvature of connecting lines (0 = straight)}

\item{label_segment_angle}{The angle for curved connectors}

\item{label_segment_ncp}{The control point density for curves}

\item{max_overlaps}{The maximum number of overlaps for the gene labels}
}
\value{
A volcano plot
}
\description{
Generate a volcano plot for gene expression analysis based on
             the log2 fold change and the -log10 p-value.
}
\details{
This function generates a volcano plot for gene expression analysis.
}
