% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rnaseq.R
\name{plt_volcano}
\alias{plt_volcano}
\title{Volcano plot for gene expression analysis}
\usage{
plt_volcano(
  data,
  x = "logFC",
  y = "FDR",
  label = rownames(data),
  selected_labels = NULL,
  p_cutoff = 0.05,
  fc_cutoff = 0,
  up_color = "firebrick4",
  down_color = "midnightblue",
  non_de_color = "grey",
  up_label = "Up-regulated",
  down_label = "Down-regulated",
  non_de_label = "Non-DE",
  number_label = TRUE,
  num_label_size = 4.5,
  x_lim = c(min(data[[x]], na.rm = TRUE) - 0.5, max(data[[x]], na.rm = TRUE) + 0.5),
  y_lim = c(0, max(-log10(data[[y]]), na.rm = TRUE) + 1),
  x_lab = bquote(~log[2] ~ "fold change"),
  y_lab = bquote(~-log[10] ~ adj.P),
  axis_lab_size = 12,
  lab_size = 4,
  bold_text = TRUE,
  legend_lab_size = 12,
  legend_lab_position = "top",
  legend_icon_size = 5,
  draw_connectors = TRUE,
  max_overlaps = 10,
  ...
)
}
\arguments{
\item{data}{A data frame containing the gene expression data}

\item{x}{The column name of the log2 fold change}

\item{y}{The column name of the P-value for use}

\item{label}{The column name of the gene names}

\item{selected_labels}{A vector of gene names to be highlighted in the plot}

\item{p_cutoff}{The p-value cutoff for significance}

\item{fc_cutoff}{The log2 fold change cutoff for significance}

\item{up_color}{The color for up-regulated genes}

\item{down_color}{The color for down-regulated genes}

\item{non_de_color}{The color for non-differentially expressed genes}

\item{up_label}{The label for up-regulated genes}

\item{down_label}{The label for down-regulated genes}

\item{non_de_label}{The label for non-differentially expressed genes}

\item{number_label}{A logical value indicating whether to display the number 
of up and down-regulated genes}

\item{num_label_size}{The size of the number labels}

\item{x_lim}{The limits for the x-axis}

\item{y_lim}{The limits for the y-axis}

\item{x_lab}{The label for the x-axis}

\item{y_lab}{The label for the y-axis}

\item{axis_lab_size}{The size of the axis labels}

\item{lab_size}{The size of the gene labels}

\item{bold_text}{A logical value indicating whether to bold the text}

\item{legend_lab_size}{The size of the legend labels}

\item{legend_lab_position}{The position of the legend labels}

\item{legend_icon_size}{The size of the legend icons}

\item{draw_connectors}{A logical value indicating whether to draw connectors}

\item{max_overlaps}{The maximum number of overlaps for the gene labels}

\item{...}{Additional arguments to be passed to the plot function}
}
\value{
A volcano plot
}
\description{
Generate a volcano plot for gene expression analysis based on
             the log2 fold change and the -log10 p-value.
}
