% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rnaseq.R
\name{rank_de_genes}
\alias{rank_de_genes}
\title{Calculate ranks of genes for differential expression analysis}
\usage{
rank_de_genes(
  de_results,
  de_method = "edgeR",
  pvalue_col = NULL,
  log2fc_col = NULL
)
}
\arguments{
\item{de_results}{A data frame containing the results of differential
expression analysis. The data frame should have at least
a column for log2 fold change and a column for (adjusted)
p-value.}

\item{de_method}{The method used for differential expression analysis.
Must be one of "edgeR" or "DESeq2". Default is "edgeR".
Use this argument to automatically set the column names
for log2 fold change and adjusted P-value. If provided value
is not "edgeR" or "DESeq2", the user must provide both the
column names for log2 fold change and P-value using the
\code{log2fc_col} and \code{pvalue_col} arguments,
respectively.}

\item{pvalue_col}{The name of the column in \code{de_results} that contains
the p-values. If not NULL, overrides the default column
name based on \code{de_method}.}

\item{log2fc_col}{The name of the column in \code{de_results} that contains
the log2 fold changes.}
}
\value{
A data frame with the addition of the following columns:
         \item{piValue}{The pi-value of the gene}
         \item{rank}{The rank of the gene based on the pi-value}
}
\description{
This function calculates ranks of genes based on the
             differential expression analysis results for RNA-seq data.
             The pi-value is calculated using the following formula:
             \deqn{pi = -log10(p) * sign(log2fc)}
             where \eqn{p} is the (adjusted) p-value and \eqn{log2fc} is the
             log2 fold change. The rank is then calculated based on the
             pi-value.
}
