% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort_colors_by_hue}
\alias{sort_colors_by_hue}
\title{Sort the Colors by Hue}
\usage{
sort_colors_by_hue(colors)
}
\arguments{
\item{colors}{A character vector of colors (hexadecimal)}
}
\value{
A character vector of colors sorted by hue
}
\description{
This function takes an input character vector of colors and
             sorts them by hue.
}
