% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_visualize.R
\name{add_dimension_arrows}
\alias{add_dimension_arrows}
\title{Add dimension arrows}
\usage{
add_dimension_arrows(
  plot,
  reduction = "umap",
  dims = NULL,
  arrow_length = 0.1,
  arrow_x_adjust_ratio = 1,
  arrow_y_adjust_ratio = 1,
  x_coord_adjust = 0,
  y_coord_adjust = 0,
  arrow_size = 0.2,
  arrow_width = 0.9,
  arrow_text_color = "black",
  arrow_type = "closed",
  text_font_size = 4,
  relative_text_dst = 0.15
)
}
\arguments{
\item{plot}{A ggplot object}

\item{reduction}{The reduction method used to generate the plot, default is
"umap", allowed values are "umap", "tsne", "pca", "spatial"}

\item{dims}{A vector of dimension names, overwrites the default dimension
names assumed from reduction method}

\item{arrow_length}{The size of the arrow, default is 0.15 (15 percent of the plot)}

\item{arrow_x_adjust_ratio}{The x-axis adjustment ratio for the arrow,
increase (> 1) and decrease (< 1) the x-axis.
Default is 1}

\item{arrow_y_adjust_ratio}{The y-axis adjustment ratio for the arrow,
increase (> 1) and decrease (< 1) the y-axis.
Default is 1}

\item{x_coord_adjust}{Horizontal adjustment of the arrow origin, default is 0}

\item{y_coord_adjust}{Vertical adjustment of the arrow origin, default is 0}

\item{arrow_size}{The size of the arrow head, default is 0.2 cm}

\item{arrow_width}{The width of the arrow, default is 0.9}

\item{arrow_text_color}{The color of the arrow and label, default is "black"}

\item{arrow_type}{The type of the arrow, default is "closed"}

\item{text_font_size}{The font size of the text, default is 4}

\item{relative_text_dst}{The relative distance of the text from the arrow,
default is 0.15}
}
\value{
a ggplot object with the dimension arrows added
}
\description{
This function adds dimension arrows to a plot at the bottom
             left corner of the plot. The arrows represent the direction of
             the dimensions in the plot. Assumes that the plot is a ggplot
             object and does not have an existing axis.
}
